refactor: delete flow node instance log v1
